[MVVM] Initial push for listings VC.
